kernel/libc: syscall: fix syscall() prototype

+++ include/libc/sys/syscall.h
@@ -38,8 +38,10 @@
 #include <sys/k_syscall.h>
 #else
 // This is defined in src/libc/syscall.asm. The maximum number of syscall
-// arguments is 5. Other arguments should be put on the stack.
-extern int64_t syscall(uint64_t id, ...);
+// arguments is 5. Other arguments should be put on the stack. We use `int`
+// types to be arch-specific, which isn't usually a good thing except when we
+// want that, like... here.
+extern int syscall(int id, ...);
 #endif
 
 /**
